refactor(runtime): reorganize factory and loader modules with utils submodules
