+++ crates/rari/src/runtime/factory/utils/js/component_loader_helper.js
