+++ crates/rari/src/runtime/factory/utils/js/metadata_collector_helper.js
