+++ crates/rari/src/runtime/module/loader/stubs/loader.rs
