+++ crates/rari/src/runtime/module/loader/stubs/react.rs
